+++ virvo/virvo/vvrenderer.cpp
@@ -268,9 +268,9 @@ float vvRenderState::getParameter(const ParameterType param) const
   case VV_QUALITY:
     return _quality;
   case VV_MIP_MODE:
-    return _mipMode;
+    return (float)_mipMode;
   case VV_ALPHA_MODE:
-    return _alphaMode;
+    return (float)_alphaMode;
   case VV_CLIP_PERIMETER:
     return _clipPerimeter;
   case VV_BOUNDARIES:
@@ -294,9 +294,9 @@ float vvRenderState::getParameter(const ParameterType param) const
   case VV_SPHERICAL_ROI:
     return _sphericalROI;
   case VV_BRICK_TEXEL_OVERLAP:
-    return _brickTexelOverlap;
+    return (float)_brickTexelOverlap;
   case VV_TEX_MEMORY_SIZE:
-    return _texMemorySize;
+    return (float)_texMemorySize;
   case VV_FPS_DISPLAY:
     return _fpsDisplay;
   case VV_GAMMA_CORRECTION:
@@ -308,7 +308,7 @@ float vvRenderState::getParameter(const ParameterType param) const
   case VV_IMAGE_SCALE:
     return _imageScale;
   case VV_IMAGE_PRECISION:
-    return _imagePrecision;
+    return (float)_imagePrecision;
   case VV_SHOW_TEXTURE:
     return _showTexture;
   case VV_OPAQUE_GEOMETRY_PRESENT:
@@ -339,9 +339,9 @@ vvVector3 vvRenderState::getParameterV3(const ParameterType param) const
   case VV_ROI_POS:
     return _roiPos;
   case VV_BRICK_SIZE:
-    return vvVector3(_brickSize[0], _brickSize[1], _brickSize[2]);
+    return vvVector3((float)_brickSize[0], (float)_brickSize[1], (float)_brickSize[2]);
   case VV_MAX_BRICK_SIZE:
-    return vvVector3(_maxBrickSize[0], _maxBrickSize[1], _maxBrickSize[2]);
+    return vvVector3((float)_maxBrickSize[0], (float)_maxBrickSize[1], (float)_maxBrickSize[2]);
   case VV_BOUND_COLOR:
     return _boundColor;
   case VV_PROBE_COLOR:
